+++ src/lib/transform.h
@@ -15,9 +15,9 @@ struct Transform {
   vec3_t scale;
 };
 
-vec3_t rotateX(vec3_t point, float rotation);
-vec3_t rotateY(vec3_t point, float rotation);
-vec3_t rotateZ(vec3_t point, float rotation);
+vec3_t rotateX(vec3_t point, double rotation);
+vec3_t rotateY(vec3_t point, double rotation);
+vec3_t rotateZ(vec3_t point, double rotation);
 
 vec3_t rotate(vec3_t point, vec3_t rotation);
