+++ src/lib/vector.h
@@ -2,20 +2,20 @@
 #define INC_3DGP_VECTOR_H
 
 typedef struct {
-    float x;
-    float y;
+    double x;
+    double y;
 } vec2_t;
 
 typedef struct {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 } vec3_t;
 
 typedef struct {
-  float width;
-  float height;
-  float depth;
+  double width;
+  double height;
+  double depth;
 } dimension_t;
 
 // TODO: Add functions to manipulate vectors 2D and 3D
