+++ src/main.c
@@ -14,7 +14,7 @@ bool is_running = false;
 vec3_t translateDir = { 0, 0, 0};
 vec3_t rotateDir = { 0, 0, 0};
 
-float previous_frame_time = 0;
+double previous_frame_time = 0;
 
 dimension_t cubeDimension = {
   .width = 1,
@@ -123,7 +123,7 @@ int main(void) {
 
     while(is_running) {
       // Fixed frame rate
-      float timeToWait = SDL_GetTicks() - previous_frame_time + TARGET_FRAME_TIME;
+      double timeToWait = SDL_GetTicks() - previous_frame_time + TARGET_FRAME_TIME;
 
       if (timeToWait > 0 && timeToWait <= TARGET_FRAME_TIME) {
         SDL_Delay(timeToWait);
